move prepare_uri
